completed sandbox implementation

+++ src/main.cpp
@@ -2,31 +2,48 @@
 #include<unistd.h>
 #include "sandbox/sandbox.h"
 
-bool firstRun()
-{
-    return access("/opt/fragarach/rootfs/bin", F_OK);
-}
 
 bool setup()
 {
+    // this is where AlpineLinux rootfs lives
     system("mkdir -p /opt/fragarach/rootfs");
+
+    // create overlay directories
+
+    // upper- writable directory, all writes go here
     system("mkdir -p /opt/fragarach/overlay/upper");
+
+    // work- internal bookeeping for kernel
     system("mkdir -p /opt/fragarach/overlay/work");
+
+    // merged- combined view of /upper and /rootfs
     system("mkdir -p /opt/fragarach/overlay/merged");
 
-    int ret = system(
-        "wget -q https://dl-cdn.alpinelinux.org/alpine/v3.18/releases/"
-        "x86_64/alpine-minirootfs-3.18.0-x86_64.tar.gz "
-        "-O /tmp/alpine.tar.gz"
-    );
+
+    // create cgroup for fragarach
+    system("mkdir -p /sys/fs/cgroup/fragarach/");
     
-    if (ret != 0) {
-        std::cerr << "Failed to download Alpine rootfs\n";
-        return false;
+    // enable memory, cpu, and pid controllers for child cgroups
+    system("echo '+memory +cpu +pids' > /sys/fs/cgroup/cgroup.subtree_control");
+    system("echo '+memory +cpu +pids' > /sys/fs/cgroup/fragarach/cgroup.subtree_control");
+
+    // download AlpineLinux rootfs if it doesn't exist
+    if(access("/opt/fragarach/rootfs/bin", F_OK))
+    {
+        int ret = system(
+            "wget -q https://dl-cdn.alpinelinux.org/alpine/v3.18/releases/"
+            "x86_64/alpine-minirootfs-3.18.0-x86_64.tar.gz "
+            "-O /tmp/alpine.tar.gz"
+        );
+        
+        if (ret != 0) {
+            std::cerr << "Failed to download Alpine rootfs\n";
+            return false;
+        }    
+        // save rootfs to /opt/fragarach/rootfs
+        system("tar -xzf /tmp/alpine.tar.gz -C /opt/fragarach/rootfs");
+        system("rm /tmp/alpine.tar.gz");
     }
-    
-    system("tar -xzf /tmp/alpine.tar.gz -C /opt/fragarach/rootfs");
-    system("rm /tmp/alpine.tar.gz");
 
     return true;
 }
@@ -39,7 +56,11 @@ int main(int argc,char* argv[])
         return 1;
     }
 
-    if(firstRun()) if(!setup()) return 1;
+    if(!setup())
+    {
+        std::cerr<<"Setup failed\n";
+        return 1;
+    }
 
 
     std::string binaryPath=argv[1];
